patches: workaround the js2c command issue with node16-win
Apparently with nodejs/node@54dfdbccc and our addition of a bootstrap
script entry, we hit some sorts of limits on Windows platform.

Mysteriously a command starts to fail. I suspect that we happen to
reach the maximum command line argument length on Windows.

This change is a quick hack to bring the command below the limit. It
does not change the functionality in any way. Upstream is going to
notice this issue soon, when they add another internal JavaScript
function.